Added documentation to Game, consolidated code from both TerminalGame and Game to Game.
